feat: QuizPackRepository 메서드 추가 및 QuizQueryRepository 메서드 추가 및 테스트 작성

+++ src/main/java/org/example/quizbox/common/presentation/Pagination.java
@@ -2,6 +2,6 @@ package org.example.quizbox.common.presentation;
 
 public record Pagination(
         long lastId,
-        long chunk,
+        int chunk,
         boolean finish
 ) {}
